// apps/dashboard/src/api/types/Unsubscribe.ts
export interface Unsubscribe_dropOut_subscription {
  __typename: "StripeSubscriptionType";
  /**
   * The ID of the object.
   */
  id: string;
  /**
   * A date in the future at which the subscription will automatically get canceled.
   */
  cancelAt: any | null;
  /**
   * If the subscription has been canceled, the date of that cancellation. If the subscription was canceled with ``cancel_at_period_end``, canceled_at will still reflect the date of the initial cancellation request, not the end of the subscription period when the subscription is automatically moved to a canceled state.
   */
  canceledAt: any | null;
  /**
   * If the subscription has been canceled with the ``at_period_end`` flag set to true, ``cancel_at_period_end`` on the subscription will be true. You can use this attribute to determine whether a subscription that has a status of active is scheduled to be canceled at the end of the current period.
   */
  cancelAtPeriodEnd: boolean;
  /**
   * The customer associated with this subscription.
   */
  customer: Unsubscribe_dropOut_subscription_customer;
  /**
   * End of the current period for which the subscription has been invoiced. At the end of this period, a new invoice will be created.
   */
  currentPeriodEnd: any;
  /**
   * Start of the current period for which the subscription has been invoiced.
   */
  currentPeriodStart: any;
  /**
   * If the subscription has ended (either because it was canceled or because the customer was switched to a subscription to a new plan), the date the subscription ended.
   */
  endedAt: any | null;
  /**
   * The subscription that this invoice was prepared for, if any.
   */
  invoices: Unsubscribe_dropOut_subscription_invoices;
  /**
   * The status of this subscription.
   */
  status: SubscriptionStatus;
}

export interface Unsubscribe_dropOut {
  __typename: "DropOutUserMutation";
  ok: boolean | null;
  error: string | null;
  subscription: Unsubscribe_dropOut_subscription | null;
  accountStatus: AccountStatus | null;
}

export interface Unsubscribe {
  dropOut: Unsubscribe_dropOut | null;
}

export interface UnsubscribeVariables {
  userId: string;
  atPeriodEnd: boolean;
}